Project Migrate Crt Lst Edt Update, OnetoMany Relationships is finished.

+++ app/Http/Controllers/AdminPanel/AdminProjectController.php
@@ -127,7 +127,9 @@ class AdminProjectController extends Controller
     public function destroy(Project $project, $id)
     {
         $data  = Project::find($id);
-        Storage::delete($data->image);
+        if ($data->image && Storage::disk('public')->exists($data->image)){
+          Storage::delete($data->image);
+        }
         $data->delete();
         return redirect('admin/project');
