+++ app/Http/Controllers/AdminPanel/CategoryController.php
@@ -135,7 +135,9 @@ class CategoryController extends Controller
     public function destroy(Category $category, $id)
     {
         $data  = Category::find($id);
-        Storage::delete($data->image);
+        if ($data->image && Storage::disk('public')->exists($data->image)){
+          Storage::delete($data->image);
+        }
         $data->delete();
         return redirect('admin/category');
